requirements.txt

cloudinary==1.40.0
Django==5.0.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-oauth-toolkit==2.4.0
django-rest-passwordreset==1.4.1
djangorestframework==3.15.1
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
PyJWT==2.8.0
whitenoise==6.6.0
python-dotenv==1.0.1
pillow==10.3.0